refactor: Move `src/utils/config` to `src/config`; Move `config/readers` to `config/utils`
