+++ src/checkers/semver-checker.ts
@@ -1,13 +1,13 @@
 import { satisfies } from "semver";
 
 import type { CliOptions } from "../command.js";
-import { SemverRuleViolation, stats } from "../stats/index.js";
 import type {
   DependencyName,
   SemverRule,
   SemverRuleset,
-} from "../utils/config/index.js";
-import { Severity } from "../utils/config/index.js";
+} from "../config/index.js";
+import { Severity } from "../config/index.js";
+import { SemverRuleViolation, stats } from "../stats/index.js";
 import type { DependenciesInstalled } from "../utils/get-dependencies-installed.js";
 import { isArrayOfArrays } from "../utils/type-guards/is-array-of-arrays.js";
