+++ src/config/utils/index.ts
@@ -1,0 +1,1 @@
+export { readDepscopConfig } from "./read-depscop-config.js";
