+++ src/config/utils/read-typescript-config.ts
