+++ src/index.ts
@@ -1,2 +1,2 @@
-export type { DepscopConfig } from "./utils/config/index.js";
-export { Severity } from "./utils/config/index.js";
+export type { DepscopConfig } from "./config/index.js";
+export { Severity } from "./config/index.js";
